refactor: move endref
