+++ src/frontend/web/src/hook/chatScroll.ts
@@ -1,4 +1,5 @@
 import { useEffect, useRef, useState } from 'react';
+import cloneDeep from 'lodash/cloneDeep';
 
 import { getOldChat } from '../Api/chat';
 import { ChatMessage } from '../../types/chat.type';
@@ -8,6 +9,8 @@ export default function chatScrollHook(
   msgList: ChatMessage[],
   setMsgList: React.Dispatch<React.SetStateAction<ChatMessage[]>>,
 ) {
+  const endRef = useRef<null | HTMLDivElement>(null);
+  const throttler = useRef(false);
   const scrollObserverRef = useRef(null);
   const [scrollLoading, setScrollLoading] = useState(false);
 
@@ -15,21 +18,39 @@ export default function chatScrollHook(
     threshold: 0.3,
   };
 
-  const getOldChatHandler = async (entries: IntersectionObserverEntry[]) => {
+  const getOldChatHandler = (entries: IntersectionObserverEntry[]) => {
     const [entry] = entries;
-    console.log(entry);
+
     if (entry.isIntersecting) {
-      const date = (entry.target as HTMLElement).dataset.date;
-      if (date) {
-        setScrollLoading(true);
-        setTimeout(() => setScrollLoading(false), 1000);
-        // const response = await getOldChat(channelId.toString(), date);
-        // console.log({ response });
-        // setMsgList(old + current)
-        console.log('hello api called');
+      if (!throttler.current) {
+        throttler.current = true;
+        setTimeout(async () => {
+          console.log('hello', entry.target); // console
+          const date = (entry.target as HTMLElement).dataset.date;
+          setScrollLoading(true);
+          const response = await getOldChat(channelId.toString(), date!);
+          if (response) {
+            const old = cloneDeep(response.content);
+            const current = cloneDeep(msgList);
+            console.log({ old, current }); // console
+            setMsgList([...old, ...current]);
+          } else console.error('fail getting old message');
+          setScrollLoading(false);
+          throttler.current = false;
+        }, 1000);
       }
     }
   };
+
+  const scrollToBottom = () =>
+    endRef.current?.scrollIntoView({
+      behavior: 'smooth',
+      block: 'nearest',
+    });
+
+  useEffect(() => {
+    scrollToBottom();
+  }, [msgList]);
 
   useEffect(() => {
     const observer = new IntersectionObserver(getOldChatHandler, option);
@@ -42,5 +63,5 @@ export default function chatScrollHook(
     };
   }, [msgList]);
 
-  return { scrollObserverRef, scrollLoading };
+  return { scrollObserverRef, scrollLoading, endRef };
 }
